flome-backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]
sqlalchemy>=2.0.0
psycopg2-binary
pydantic>=2.0.0
pydantic-settings
python-dotenv
python-multipart
